Only use the properties of the original document, not its versions
Now that we are able to see multiple versions of a document, we do not want
to check the properties of the *versions* of a document to see if it is
published or not - we only want to check the properties of the original
document. The versions do not have properties.

Amend the `get_judgment_xml` and `eval_xslt` methods and XQueries to take a
verion uri as well as a judgment uri, and only use the judgment uri to
query properties. If a version uri is passed in, then assume that is the
document intended to be returned.