Replace lxml with xml.etree from the Python standard library
LXML requires the system it runs on to have lxml binaries available, which is not always the case (for example when executing in a lambda environment).

As none of the functionality we use is lxml-specific, we can use the xml library that comes as part of the standard python library, guaranteeing that it will be available in all environments without requiring extra build steps.