If Elements are missing from the XML, send empty ones
If Elements are missing from the XML which we need to edit (e.g
if we are attempting to edit the metadata name of a judgment but its
`akn:FRBRWork` element is completely missing) then send an empty Element
in its place instead of an error.

This will allow editors to correct the missing metadata.