Set judgment metadata name in the API Client
We are currently editing and saving the Judgment XML in the Editor UI project,
which is creating some issues with namespaces being removed from the judgment
XML.

This method allows the metadata name to be set directly using XQuery and does
not require any serialising or de-serialising of the XML on the editor UI side.

TODO: Refactor these to be more generic. The issue here is that passing in an
XPath to the node to be updated as a string does not work, as XQuery expects
the XPath to be a node()*, not a string, and I'm not sure how to cast a
string to a node()*, or how to pass the Xpath in as a node()*