Set metadata court and neutral citation via XQuery
Update the values of these nodes directly via XQuery.

TODO: Refactor these to be more generic. The issue here is that passing in an
XPath to the node to be updated as a string does not work, as XQuery expects
the XPath to be a node()*, not a string, and I'm not sure how to cast a
string to a node()*, or how to pass the Xpath in as a node()*