Fix: If metadata name element does not exist, insert a new one as a child
We were assuming that the `akn:FRBRname` node would exist and could be edited,
but in some cases this node does not exist if the parser was unable to find a
case name in the judgment.

Check to see if a `akn:FRBRname` node exists in the document, and if it does
replace it, otherwise create a new one.