If the `uk:court` element does not exist, create a new one
We have had reports of the court metadata on judgments not saving. This is
due to the `uk:court` element being entirely missing, rather than existing but
being empty. If the element is entirely missing ,the original XQuery to replace
and existing node does not work - we need to create an entirely new node.