Set the judgment's internal URI (`FRBRthis` and `FRBRuri` nodes)
In https://github.com/nationalarchives/ds-caselaw-public-access-service/blob/main/marklogic/src/main/ml-modules/root/judgments/xslts/judgment2.xsl#L69
we can see that the XSLT uses the node at `/akomaNtoso/judgment/meta/identification/FRBRWork/FRBRthis/@value`
to determine the URI of a document, and then uses that URI to build up a URL
for the judgment's images.

In judgments which have failed processing, and have had a Neutral Citation Number
(and therefore a URI) added manually by the editors, this internal URI at
`/FRBRWork/FRBRthis/@value` remains blank.

This has the knock on effect of image URLs not resolving correctly in the HTML
version of a judgment (which is generated from the above XSLT).

This commit introduces an XQuery & API endpoint to set the value of
`FRBRWork/FRBRthis/@value`

This can then be used in the Editor UI to set the internal URI of a judgment
once a judgment is edited by the editors.

Additionally, the values of FRBRExpression/FRBRthis/@value and FRBRExpression/FRBRuri/@value
should be set to the URI (but without the `/id/`), and the values of
FRBRManifestation/FRBRthis/@value and FRBRManifestation/FRBRuri/@value should be set
to the URI, without the `/id/` and with `data.xml` on the end.

E.g.

```
<identification source="#tna">
  <FRBRWork>
    <FRBRthis value="https://caselaw.nationalarchives.gov.uk/id/ewhc/scco/2022/1148" />
    <FRBRuri value="https://caselaw.nationalarchives.gov.uk/id/ewhc/scco/2022/1148" />
    <FRBRdate date="2022-05-04" name="judgment" />
    <FRBRauthor href="#ewhc-seniorcourtscosts" />
    <FRBRcountry value="GB-UKM" />
    <FRBRnumber value="1148" />
  </FRBRWork>
  <FRBRExpression>
    <FRBRthis value="https://caselaw.nationalarchives.gov.uk/ewhc/scco/2022/1148" />
    <FRBRuri value="https://caselaw.nationalarchives.gov.uk/ewhc/scco/2022/1148" />
    <FRBRdate date="2022-05-04" name="judgment" />
    <FRBRauthor href="#ewhc-seniorcourtscosts" />
    <FRBRlanguage language="eng" />
  </FRBRExpression>
  <FRBRManifestation>
    <FRBRthis value="https://caselaw.nationalarchives.gov.uk/ewhc/scco/2022/1148/data.xml" />
    <FRBRuri value="https://caselaw.nationalarchives.gov.uk/ewhc/scco/2022/1148/data.xml" />
    <FRBRdate date="2022-06-16T00:40:44" name="transform" />
    <FRBRauthor href="#tna" />
    <FRBRformat value="application/xml" />
  </FRBRManifestation>
</identification>
```