Respect & maintain the `FRBRdate` attribute `name` in `FRBRWork`
Related to https://github.com/nationalarchives/ds-caselaw-public-ui/pull/316

This XQuery sets or updates the node at `/akn:akomaNtoso/akn:judgment/akn:meta/akn:identification/akn:FRBRWork/akn:FRBRdate`

However, it does not respect the value of the `@name` attribute. It is being
set to `judgment`, which is true in many cases, but may also be `hearing` in
some cases.

In the referred PR above, we changed the way we get a judgment's date, from
looking for an `FRBRdate` with an attribute of `judgment`, we instead
look for any `FRBRdate` inside the `FRBRWork` element. So the attribute `name`
value is redundant.

When updating a judgement's date, we do not want to obliterate the existing
`name` attribute value, and replace it with `judgment`, as we were previously
doing. Instead we want to preserve it in the new/updated `FRBRWork/FRBRdate`
element.

However, if a judgment does not have an `FRBRWork/FRBRdate` element, create
a new one and give it the `@name` attribute `judgment` by default.