Add methods to get the lock/checkout status of a judgment
In the privileged API, we want to be able to see the checkout/lock status of a
judgment.

`get_judgment_checkout_status.xqy` checks the checkout status of a judgment using
`dls:document-checkout-status($uri)`

The result will either be empty, if the judgment is not locked, or an XML
fragment containing the URI and annotation message, if the judgment is locked.

The method `get_judgment_checkout_status` can be called on the Client, and will
return a response with either an XML fragment describing who has locked the
judgment, or empty if it is not locked.

The method `get_judgment_checkout_status_message` can also be called on the
Client, and will return either the lock annotation, if the judgment is locked,
or "not locked" if the judgment is not locked.