Add `user_has_privilege` method & XQuery to check if a user has a privilege
A user has "privileges" in Marklogic, which are represented by names & IDs.
When we want to check to see if a user has a certain privilege, we can get the
ID of the privilege we want to check, then a sequence of the privileges a user
has, and see if the former is in the latter.

The XQuery returns `true` or `false`. `fn:boolean` will return true for any
positive integer value, and false for 0, an empty sequence or an empty string.

The method returns a response object
with a `text` value of `true` or `false`.

If the requested privilege does not exist, Marklogic will throw `SEC-PRIVDNE`
- map this error onto `MarklogicNotPermittedError` to indicate you have tried
to do something not permitted.