Use `user_has_privilege()` to check if a user can see unpublished documents
In https://github.com/nationalarchives/ds-caselaw-public-access-service/pull/65
we propose to add a custom privilege which indicates a user is allowed to view
unpublished judgments.

This commit uses `user_has_privilege()` to check that a user has this privilege
(which is found via its unique URI and action). The method returns `True` or
`False`.