Ensure only people who are allowed to view unpublished judgments can view them
We have added `user_can_view_unpublished_judgments(self.username)` to the
client, which allows us to check that the currently logged in user (`username`)
has the appropriate permission to view unpublished judgments.

If a user attempts to access unpublished judgments by (somehow) calling an
endpoint with `show_unpublished=True` BUT they do not have the correct privilege
to view unpublished judgments, they should not be allowed to do that. Instead
`show_unpublished` will be set to False and they will not be able to see
unpublished judgments.

If a user who *is* allowed to view unpublished judgments asks specifically to
view unpublished judgments, by passing in `show_unpublished=False`, the value
of `show_unpublished` will remain False to allow them to view the required
unpublished judgments.