Refactor: Handle viewing of unpublished judgments in a separate function
Refactor out the logic which determines if a user is allowed to view unpublished
judgments, and if they are asking to view unpublished judgments.

If they are 1) not permitted to view unpublished judgments, and 2) asking to
view unpublished judgments, do not allow them AND log a warning that the user
is attempting to do something they're not permitted to do.

If the user CANNOT view unpublished judgments, and attempts to view unpublished
judgments, `show_unpublished` is FALSE and a warning is logged
If the user CANNOT view unpublished judgments, and does not attempt to view
unpublished judgments, `show_unpublished` is FALSE
If the user CAN view unpublished judgments, and attempts to view unpublished
judgments, `show_unpublished` is TRUE
If the user CAN view unpublished judgments, but specifically asks not to view
unpublished judgments, `show_unpublished` is FALSE