Break judgment checkout
Add the ability to break a checkout lock on a judgment. If a judgment is locked
we need a way to break the lock and return it to an editable state.

The XQuery uses `dls:break-checkout($uri, $deep)` to break the lock and
any locks on associated versions.

The function `break_checkout()` returns a request object containing no content
- the XQuery returns an empty sequence (assuming the requested URI exists).