Add deprecation warnings for GET() and POST()
We suspect these methods are no longer used in the client, and have been
replaced with eval() and invoke(). Add a warning to these methods so we can
check they are definitely not being used (e.g. by checking staging logs after
this version of the API Client is deployed).