Change assertions on tests which call the `eval` method
Change the assertions on tests which call eval().

Previously we were always
testing that the `accept_header="application/xml"`, which is always true
and needs only to be tested once (we have retained this assertion in
`test_eval_xslt`). Remove the repeated assertion to reduce noise.

Change the other assertions to look for named `kwargs` and `args` for clarity
and brevity.