Allow users with the `admin` role to view unpublished judgments
Allow users who have the `admin` role to view unpublished judgments.
This feels slightly backwards to me but unfortunately the `admin` role is not
editable in Marklogic, so granting the `admin` user the correct custom privilege
to view unpublished judgments is not immediately obvious.

As we have now figured out our custom roles in a previous PR (
https://github.com/nationalarchives/ds-caselaw-public-access-service/pull/86)
this work seems less urgent (as we can tailor our custom roles and should not
need to grant admin access to users). But it is still useful to be able to view
unpublished judgments as an admin user, for completeness.

There may be a way to grant this access via Marklogic's QBAC system
https://docs.marklogic.com/guide/security/qbac - for future investigation.