Ensure Work Date and Court values are returned as text
In https://github.com/nationalarchives/ds-caselaw-custom-api-client/pull/94
we changed the output of `get_judgment_citation` to be the text value of the
citation, not the response. This Commit beings `get_judgment_court` and
`get_judgment_work_date` in line with that function.

These functions are not yet used anywhere so it's safe to amend them.