Move Judgment object from EUI
This is a fairly blind copying of the Judgment model and its associated
tests, so we can begin the process of decoupling EUI from its existing
implementation and toward using this. Some methods haven't been copied
because they are (at least for the moment) strongly coupled to the way
EUI understands judgments.

The lack of tests for AWS is a known problem, but outside the scope of
this PR.