Add strict typing for dicts passed to MarkLogic
This introduces a new script, `build_xquery_type_dicts`, which looks
through the `.xqy` files in `src/caselawclient/xquery` and uses their
contents to generate a new `xquery_type_dicts.py`. This file strictly
declares all required and optional parameters for each query as
`TypedDict`s, which in this case are subclassed via `MarkLogicAPIDict`.

`Client.py` requires that all `dict`s passed to `_send_to_eval` are
instances of `MarkLogicAPIDict`, which means they will fail if they are
not one of the `TypedDict`s generated by the script.

As a final belt-and-braces, generating these typings is added to
pre-commit; changing the parameters in a `.xqy` file will cause changes
to be made in `xquery_type_dicts.py`, which will in turn cause mypy to
fail if they are incompatible with the current implementation.