Judgment now correctly raises an exception if URI doesn't exist
`Judgment` originally looked for a neutral citation to test for the
presence of the judgment document, but this function would return an
empty string rather than the (expected) `MarklogicResourceNotFoundError`
exception if the document didn't exist.

This implements a new `judgment_exists` method to explicitly check for
the presence of a document in MarkLogic, and re-tools the
`Judgment.__init__` method to use this instead of using the NCN as a
proxy.

This is a breaking change, because the exception thrown by a `Judgment`
if it isn't found is now `JudgmentNotFound` rather than
`MarklogicResourceNotFoundError`.

Co-authored-by: Dragon <david.mckee@dxw.com>