Move simple getters for Judgment into a data structure
Since a lot of what the `Judgment` class does is very similar, we can
start to coalesce it into simpler functions which behave dynamically
based on a data structure.

In this case, getters for attributes which only call a method on the API
client accepting a single URI as a parameter are now defined in this way
and make use of the `__getattr__` dunder method.

There is an internal implementation of caching values in `values_store`,
as a precursor to the end goal of this work where we can initialise a
whole bunch of these values along with the class.