Add detection of 'parked' URIs
Where the ingester detects that it cannot assign a URI to a document (for example, the document is claiming to have the same URI as an existing document and it needs manual version reconciliation) then it will assign a URI in the format `parked/{uuid}`.

This PR adds detection of those parked URIs to the API client; parked URIs cannot be published.