Add new get_document_by_uri method to MarklogicAPIClient
This (currently) returns only a Judgment, since this is the only
possible document class.

This means that all downstream users can swap to using
`MarklogicAPIClient.get_document_by_uri` with an existing instance of
`MarklogicAPIClient` to retrieve a document of any type without knowing
its type in advance. In future, this method will implement type
detection logic to make sure it's initialising the correct document
class.