Improve handling of multipart responses
This entire client class currently assumes that there is exactly one
item coming from MarkLogic in a multipart response, or failing that the
only useful part of the response is the first item. This is currently
true, but will not necessarily be in the future (and is prone to
flakiness in any case).

This moves handling of the multipart response and turning it into text
into a distinct function which can be used in future to retrieve a list
of results, and simplifies the existing `decode_multipart` function (to
be renamed later) to return only the first result, or fail with a hard
exception (because it would mean whatever is calling the method is
expecting a single result in a context where multiple results have been
returned).

This should _not_ be a breaking change if downstream consumers are using
the client in the intended way, but we can't guarantee it.