Fix poetry installation for docs generation
This was only installing pdoc and its dependencies, which is everything
needed to generate the docs, but did not include things needed to
actually end up with valid code for the documentation generator to
parse.

These dependencies are now in an optional group so they're not installed
automatically (eg we have no need for them in production, and most
people have no need for them in dev).