Ensure marklogic responses are treated as unencoded bytes
Marklogic returns strings with an encoding declaration, which are unsupported by lxml (which gives precedence to the encoding declaration in the xml itself), and raises an ValueError with message "Unicode strings with encoding declaration are not supported. Please use bytes input or XML fragments without declaration." - this can be verified by attempting to access any judgment on staging markdown with version 13.0.0 of the client.

Instead, we can explicitly convert the marklogic response to unencoded bytes before parsing with lxml (see this stack overflow answer: https://stackoverflow.com/questions/28534460/lxml-etree-xml-valueerror-for-unicode-string)  - this PR does that, following the same convention used in `models.utilities.get_judgment_root` and adds a test for the desired behaviour.