Add new `Document.delete()` method
Running this against an instance of a `Document` (or subclass like
`Judgment` or `PressSummary`) will remove the document from MarkLogic
and S3.

A document _cannot_ be deleted if it is published, since this implies it
as already been given a publicly accessible URI. A document must first
be unpublished (hopefully triggering appropriate workflows for hadnling
depublication with data reusers etc) before it can be deleted.