Be more strict about the flavours of strings
There is currently ambiguity about the types of strings which are being
passed around, particularly with the overloaded term "URI". This uses
`typing.NewType` to create two new types of string, `DocumentURIString`
and `MarkLogicDocumentNameString`, representing "a bare URI for a
document's identifier" and "the exact name of a document's
representation in MarkLogic" respectively.

This adds to our existing type checking in making sure that we're
passing the right type of string between our various functions.