Change `is_failure` behaviour to rely on `failed_to_parse`
The `is_failure` parameter previously relied on checking the URI to
determine if a failure state existed. This is a hangover from previous
behaviour, but not one we want.

Change the behaviour so that it instead relies on checking other
parameters which are indicative of a failure. At the moment this is only
`failed_to_parse`, but this list may be expanded in the future.