Expose how the latest version came into being.
a new `document.get_latest_manifestation_type` method which returns either "tna-enrich" or "transform" depending on the "name" property of the most recent manifestation date.

This is used in the versions table in the editor UI to annotate each version with it's 'version type', which helps editors better understand where each version came from.