Make annotations mandatory instances of VersionAnnotation
All functions which add an annotation to a version of a document now
require an annotation to be provided, and this must be an instane of
`VersionAnnotation`. This enforces a structure to the annotation,
outputting a JSON string which will be saved to the database, ready for
unpacking elsewhere.

There are enumerated types of versions, indicating the reason for the
version being created, in the `VersionType` enum.

Functions calling `VersionAnnotation.to_json` must first use
`VersionAnnotation.set_calling_function` to record their own name. This
won't be exposed in the UI in most cases, but is recorded for debugging
purposes.