Highlight search term in judgment
`document.content_as_html` now takes an optional `query` parameter which is preprocessed in the same manner as the search query, then used to highlight matching terms in the document, which are returned with a surrounding `<mark>` tag. This mark tag has an id which gives the number, in sequence of the match from 0..n in the form `mark_{x}`.

A new `document.number_of_mentions` method which takes a `query` parameter and returns the number of highlighted mentions returned in the html for the query