Move creation of `SearchResultMetadata` into `SearchResult`
Previously, `SearchResultMetadata` had a static `create_from_uri` method
which was being called by `SearchResult`, which then ultimately
initialised an instance of `SearchResultMetadata`.

However, this is the _only_ place where this method was being called
from. This means we can push this method upwards into `SearchResult` and
avoid the need for a class acting as its own factory.