Refactor use of static api_client out of search
Search, specifically metadata handling, was relying on the static
instance of api_client which was initialised as part of loading this API
library. This instance is due to be deprecated, which means that instead
the dynamic instance which is initialised by the calling code must be
passed back down the stack.