Stop swallowing malformed XML
The code to extract the root element of the XML is also responsible for
identifying where a document is not XML (which is theoretically
possible; MarkLogic can also store JSON, blobs and text). At the moment
if this happens the fact it's not XML is silently swallowed and we
pretend the root element is `error`, the same as a parser failure.

This behaviour could potentially mask other problems; if we encounter a
parser error then although it's unwanted it's a defined condition which
we want to be able to handle within the usual user flows. Encountering a
non-XML document means something has gone catastrophically wrong, and we
want things to fail hard and shout about it.

Signed-off-by: David McKee <david.mckee@dxw.com>