Create new `Document.XML` subclass
This new subclass is intended to contain all of the functions needed to
interact with a document's XML representation, to keep them isolated and
break up the class somewhat.

To begin with, all it does is store a bytestring representation of the
document (as it comes back from the database) and provides methods to
interact with it as a string and a tree representation.