Document XML as string is now derived from the XML tree
Instead of storing multiple representations of the XML, we can just use
`etree.tostring()` to convert the tree representation of XML back into a
bytestring, then decode it to a UTF-8 string.

This means in future we never have to worry about if we're manipulating
the text representation in line with the tree, because it's generated
from the tree every time it's called.