Perform XML conversion to tree on initialisation
The `Document.XML` class now converts the bytestring representation of
the XML to an lxml tree on initialisation, rather than relying on having
a separate method for it.

This means we can entirely ditch storing the bytestring representation,
as we can convert a tree to bytestrings on the fly if needed. It also
means we can catch cases where the document isn't valid XML at init
time, instead of waiting for a read operation.

To make this easily testable, we moved the `mock_api_client` test
fixture to `conftest.py` so it is available for all tests on-demand,
which gives us only one place to stub out test behaviours such as this
one (ie always relying on having a valid XML string). This will be
further refactored in future PRs to unify its behaviour with
`DocumentFactory` as part of creating things which behave more like real
`Document`s.