Prioritise quoted phrases in search snippets.
This goes with a [marklogic PR](https://github.com/nationalarchives/ds-caselaw-marklogic/pull/50) which ensures that snippets containing quoted phrases in the query are given priority when they're displayed to users.

This PR provides the list of quoted phrases to marklogic (essentially impossible to do with marklogic's own regex engine), and also limits the number of matched phrases returned to 4 (the current default), as the number returned by marklogic itself has necessarily increased (we need to generate a larger amount of matches before prioritisation to ensure we catch any mentions of the quoted phrase).