Add methods to get and set judgment jurisdiction
The General Regulatory Chamber has a number of Jurisdictions (or
'sub-tribunals' as we sometimes call them). These are different to the
separate Lists we see within eg the Chancery Division - Lists are an
administrative distinction, decided on by the court themselves, whereas
Jurisdictions are a statutory one, decided in law. Neither jurisdiction
nor list affects the NCN of the judgment, but they're stored differently
in the XML - Lists are represented in the court code in the uk:court
element, whereas judgments are represented with the uk:jurisdiction
element in addition to the uk:court.

This PR provides accessors for the jurisdiction data, both in isolation,
and combined with the court as a slash-separated string, following the
convention established in the courts metedata in ds-caselaw-utils.

This will allow us to display the jurisdiction in the PUI (and allow
editors to correct it in EUI) with minimal changes