Add a helper to get press summaries for a document
This adds a `get_press_summaries_for_document_uri` which gets press
summaries for a document URI. I’ve made the query itself configurable so
we can use it for other components too, but for now we’re only using it
for press summaries.

As I wasn’t super confident it was going to work, I’ve also added a
smoketest, which I think it worth having around.