Get linked document
This adds the logic to get a linked document to the individual
document class. At the moment this is done in the Public UI, which

a) Is hard to mock; and
b) Gives us more to do at the UI end

Keeping the responsibility in the client makes the behaviour more
predictable, as well as means we have a “real” document object
to get things like the title etc from.

I’ve had to do some jiggery pokery to avoid cyclic imports and return
the correct document type, but the tests seem happy