feat(FCL-268): break functions which rely on the document body into their own subclass
This change helps us better isolate behaviours which rely on the XML of the body document, making it clearer which information is coming from where. This in turn lets us improve concern boundaries, so the code is more maintainable, reliable and testable.

BREAKING CHANGE: Multiple methods which used to be within `Document` are now in `Document.body`