test(FCL-331): retool test factory behaviour
Test factories now live within the package proper (so they can be used by downstream consumers for consistent test behaviour), and now also provide real instances of the Document class instead of a mock. This gives more parity in behaviour, as well as reducing mock boilerplate.

nb we mock the methods around getting HTML because otherwise they invoke a real call to MarkLogic; they don't perform an in-app transformation. We intend to change this behaviour at some point in the future, but for now it's safest to mock.