feat: Validate strings when creating a new DocumentURIString
This improves the robustness of `DocumentURIString`` handling by ensuring that they conform to the expected rules (no trailing or leading slashes, no dots). Since initialising a document now requires an instance of this class, and we can guarantee the string contained within meets our requirements, we can safely lose string sanitisation in the `Document` class.

As part of this, `Judgment` and `PressSummary` have had their signatures updated to reinforce the expectation of `DocumentURIString` being provided as the `uri` argument. This led to identifying unsafe behaviour around the handling of press summary URIs and version URIs, which have now been made more explicit.

BREAKING CHANGE: Code which provided unsanitised URIs when initialising `DocumentURIStrings` will now cause `InvalidDocumentURIException`s to be raised.