feat(FCL-309): add stub for defining identifier schemas, and a Neutral Citation schema
In this framework both the schema defining an identifier and the identifiers themselves are implemented as abstract classes. This is because we know that some types of identifier (such as Neutral Citations) may have places in the codebase where we require additional type-specific functionality. By using type-specific concretions of those abstract classes we can guarantee we're doing so in a type-safe manner, whilst still keeping a generic framework for the underlying storage.