feat(FCL-533): modify human identifier to rely on identifiers framework
This is instead of assuming that an NCN will always be present. As a result, this has had knock-on changes in some other places which were previously relying on this incorrect behaviour. This includes some breaking changes to method signatures, which can now return a `None` in places they previously could not. Since relying on the presence of an NCN is now considered dangerous, these methods have been marked as deprecated.

BREAKING CHANGE: Methods which were previously guaranteed to return a Neutral Citation may now return `None`.