Remove query mark tags from meta section of XML
When a search query occurs for a string that occurs in the upper part
of the XML, it would be corrupted, e.g.

<uk:cite>[2024] <uk:mark id="mark_1">UKSC</uk:mark> 1</uk:cite>

if the search query included a word from the contained string.

This removes all such tags, leaving a gap if there was a match, so
this might break the JS which expects a mark_1 to exist.