refactor(Identifiers): rename Identifiers to IdentifierCollection and move to submodule
We move Identifiers into a new submodule to help us resolve some cyclic imports with future work. As part of this, since it's a breaking change anyway, we take the opportunity to rename the class so it's clearer what it is.

BREAKING CHANGE: `identifiers.Identifiers` is now `identifiers.collection.IdentifierCollection`