feat(Document): first_published_datetime now interprets sentinel values
A sentinel value of `1970-01-01 00:00`` indicates that we know a document has previously been published, but not exactly when. `has_ever_been_published` already interprets this correctly, but we add a new test for this.

We also add `first_published_datetime_display` which will return a `None` where the value matches this sentinel, and this can be used in combination with `has_ever_been_published` to indicate an "unknown" state.